Added routing between index page and property pages; Changed use of apartment to property
