+++ app.vue
@@ -2,6 +2,7 @@
 
 <template>
   <Spritesheet />
+  <NuxtLoadingIndicator color="#99f6e4" />
   <Navbar />
   <NuxtPage />
   <!-- <div class="h-[480px]"></div> -->
